ml/requirements.txt
backports.ssl-match-hostname==3.4.0.2
certifi==2015.4.28
gevent==1.0.1
gnureadline==6.3.3
greenlet==0.4.7
grequests==0.2.0
ipython==3.1.0
Jinja2==2.7.3
jsonschema==2.4.0
MarkupSafe==0.23
matplotlib==1.4.3
mistune==0.5.1
mock==1.0.1
nose==1.3.6
numpy==1.9.2
pandas==0.16.1
ptyprocess==0.5
Pygments==2.0.2
pyparsing==2.0.3
python-dateutil==2.4.2
pytz==2015.4
pyzmq==14.6.0
requests==2.7.0
scikit-learn==0.16.1
scipy==0.15.1
six==1.9.0
sklearn-pandas==0.0.9
terminado==0.5
tornado==4.1
